Add as-is class
Text is supposed to be transformed lowercase but there are some terms
that would need to look as they were written. For most of them, it does
not really matter whether they are Capitalized, UPPERCASE or lowercase.
But for some of them, it can affect legibility (AI, etc.).

By adding the class `as-is`, anyone will be able to keep the case of
a texts prepared to be rendered as HTML.