Fix button-pill height and background
This commit fixes two things:

Height
=====

Currently, the height was 4px over the documented height. This was
because for centering the text we are using the line-height trick with
the final height, this adds two times the border-width to the height.

Background color
========

The background color needs to be transparent and not inherit. They
behave the same as long as the background-color does not contain
opacity.