New grid
New grid is coming. It is actually here.

Previous grid was based on pixel units for columns and margins for
adding gutter. The new proposed grid had % units for columns and padding
used for gutter.

Making both systems somehow meet was hard, and this is the simplest
solution I could came with. In the way, I have tried a few and, while
this one is not perfect (it is not, believe me).

The current solution is a bit nasty. What it does is mixing both
measures in a special way. While column width is %-ish, gutter is fixed.
Therefore, a column width is `(100% - $total-gutter-size)
/ $num-columns`. A multi-column width, contains a percent based size
(the column width) and a pixel-based width (the gutter space).

Oh, god, I feel naughty writing that.

Why are the formulas so complicated?
------

It might be thought that the formulas could be easily simplified (and
maybe they can!) but the thing is that, when mixing units,
simplifications might not work as expected. Counter-intuitive but it
does so :$

Where did you get the breakpoints and mixins?
------

Twitter Bootstrap v4. While I don't like bootstrap for the classes they
use and a few other things, the use they do for mixins is really nice
(IMHO). They use them for asserts (they check that the breakpoints is
a valid map, for example), for adding media queries (I took
`media-breakpoint-up`, but also have the `-down` and `-between` version
that let's user abstract away from media queries (not need to use
variables in the media-queries, but abstracts away all the code).

So... this PR is not fully complete?
------

Nope, it lacks some extra mixins for media-queries (easily find in
Bootstrap v4) and fixing a couple of things...

Does this breaks anything?
------

Yes, there is a couple of things broken, mainly, the horizontal slack
form in projects section.

Other sections are slightly broken but mainly because previous version
contained two different approaches to get the same result in different
parts.

This will be fixed next.